changed controllers package structure
