+++ src/main/java/vlad/petrovskyi/internetshop/controllers/orders/AllOrdersController.java
@@ -1,4 +1,4 @@
-package vlad.petrovskyi.internetshop.controllers.db.orders;
+package vlad.petrovskyi.internetshop.controllers.orders;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -8,14 +8,14 @@ import javax.servlet.http.HttpServletResponse;
 import vlad.petrovskyi.internetshop.lib.Injector;
 import vlad.petrovskyi.internetshop.service.OrderService;
 
-public class DeleteOrderController extends HttpServlet {
+public class AllOrdersController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("vlad.petrovskyi.internetshop");
     private final OrderService orderService =
             (OrderService) INJECTOR.getInstance(OrderService.class);
 
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        orderService.delete(Long.parseLong(req.getParameter("order_id")));
-        resp.sendRedirect(req.getContextPath() + "/getAllOrders");
+        req.setAttribute("orders", orderService.getAll());
+        req.getRequestDispatcher("/WEB-INF/views/orders/allOrders.jsp").forward(req, resp);
     }
 }
