+++ src/main/java/vlad/petrovskyi/internetshop/controllers/orders/CompleteOrderController.java
@@ -1,4 +1,4 @@
-package vlad.petrovskyi.internetshop.controllers.user;
+package vlad.petrovskyi.internetshop.controllers.orders;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
