+++ src/main/java/vlad/petrovskyi/internetshop/controllers/orders/DeleteOrderController.java
@@ -1,4 +1,4 @@
-package vlad.petrovskyi.internetshop.controllers.user;
+package vlad.petrovskyi.internetshop.controllers.orders;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -7,19 +7,15 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import vlad.petrovskyi.internetshop.lib.Injector;
 import vlad.petrovskyi.internetshop.service.OrderService;
-import vlad.petrovskyi.internetshop.service.UserService;
 
-public class UserOrdersController extends HttpServlet {
-    private static final String USER_ID = "user_id";
+public class DeleteOrderController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("vlad.petrovskyi.internetshop");
     private final OrderService orderService =
             (OrderService) INJECTOR.getInstance(OrderService.class);
-    private final UserService userService = (UserService) INJECTOR.getInstance(UserService.class);
 
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        req.setAttribute("orders", orderService.getUserOrders(
-                userService.get((Long) req.getSession().getAttribute(USER_ID))));
-        req.getRequestDispatcher("/WEB-INF/views/orders/userOrders.jsp").forward(req, resp);
+        orderService.delete(Long.parseLong(req.getParameter("order_id")));
+        resp.sendRedirect(req.getContextPath() + "/getAllOrders");
     }
 }
