+++ src/main/java/vlad/petrovskyi/internetshop/controllers/orders/UserOrdersController.java
@@ -1,4 +1,4 @@
-package vlad.petrovskyi.internetshop.controllers.db.users;
+package vlad.petrovskyi.internetshop.controllers.orders;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -6,19 +6,20 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import vlad.petrovskyi.internetshop.lib.Injector;
+import vlad.petrovskyi.internetshop.service.OrderService;
 import vlad.petrovskyi.internetshop.service.UserService;
 
-public class DeleteUserController extends HttpServlet {
-
+public class UserOrdersController extends HttpServlet {
+    private static final String USER_ID = "user_id";
     private static final Injector INJECTOR = Injector.getInstance("vlad.petrovskyi.internetshop");
+    private final OrderService orderService =
+            (OrderService) INJECTOR.getInstance(OrderService.class);
     private final UserService userService = (UserService) INJECTOR.getInstance(UserService.class);
 
-    @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        String userId = req.getParameter("user_id");
-        userService.delete(Long.valueOf(userId));
-
-        resp.sendRedirect(req.getContextPath() + "/getAllUsers");
+        req.setAttribute("orders", orderService.getUserOrders(
+                userService.get((Long) req.getSession().getAttribute(USER_ID))));
+        req.getRequestDispatcher("/WEB-INF/views/orders/userOrders.jsp").forward(req, resp);
     }
 }
