+++ src/main/java/vlad/petrovskyi/internetshop/controllers/orders/ViewOrderController.java
@@ -1,4 +1,4 @@
-package vlad.petrovskyi.internetshop.controllers.db.orders;
+package vlad.petrovskyi.internetshop.controllers.orders;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
