+++ src/main/java/vlad/petrovskyi/internetshop/controllers/products/AddProductController.java
@@ -1,14 +1,16 @@
-package vlad.petrovskyi.internetshop.controllers.db.products;
+package vlad.petrovskyi.internetshop.controllers.products;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import vlad.petrovskyi.internetshop.lib.Injector;
+import vlad.petrovskyi.internetshop.model.Product;
 import vlad.petrovskyi.internetshop.service.ProductService;
 
-public class DeleteProductController extends HttpServlet {
+public class AddProductController extends HttpServlet {
 
     private static final Injector INJECTOR = Injector.getInstance("vlad.petrovskyi.internetshop");
     private final ProductService productService =
@@ -17,7 +19,16 @@ public class DeleteProductController extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        productService.delete(Long.valueOf(req.getParameter("product_id")));
+        req.getRequestDispatcher("/WEB-INF/views/products/addNewProduct.jsp").forward(req, resp);
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp)
+            throws ServletException, IOException {
+        String name = req.getParameter("name");
+        String price = req.getParameter("price");
+
+        productService.create(new Product(name, BigDecimal.valueOf(Long.parseLong(price))));
         resp.sendRedirect(req.getContextPath() + "/allProductsFromDb");
     }
 }
