+++ src/main/java/vlad/petrovskyi/internetshop/controllers/products/GetAllProductsInDbController.java
@@ -1,4 +1,4 @@
-package vlad.petrovskyi.internetshop.controllers.user;
+package vlad.petrovskyi.internetshop.controllers.products;
 
 import java.io.IOException;
 import java.util.List;
@@ -10,7 +10,7 @@ import vlad.petrovskyi.internetshop.lib.Injector;
 import vlad.petrovskyi.internetshop.model.Product;
 import vlad.petrovskyi.internetshop.service.ProductService;
 
-public class GetAllProductsController extends HttpServlet {
+public class GetAllProductsInDbController extends HttpServlet {
 
     private static final Injector INJECTOR = Injector.getInstance("vlad.petrovskyi.internetshop");
     private final ProductService productService =
@@ -20,7 +20,8 @@ public class GetAllProductsController extends HttpServlet {
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
         List<Product> productList = productService.getAll();
+
         req.setAttribute("products", productList);
-        req.getRequestDispatcher("WEB-INF/views/products/allProductsBuy.jsp").forward(req, resp);
+        req.getRequestDispatcher("/WEB-INF/views/products/allProductsEdit.jsp").forward(req, resp);
     }
 }
