+++ src/main/java/vlad/petrovskyi/internetshop/controllers/products/GetAllProductsOfUserController.java
@@ -1,7 +1,7 @@
-package vlad.petrovskyi.internetshop.controllers.db.products;
+package vlad.petrovskyi.internetshop.controllers.products;
 
 import java.io.IOException;
-import java.math.BigDecimal;
+import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -10,7 +10,7 @@ import vlad.petrovskyi.internetshop.lib.Injector;
 import vlad.petrovskyi.internetshop.model.Product;
 import vlad.petrovskyi.internetshop.service.ProductService;
 
-public class AddProductController extends HttpServlet {
+public class GetAllProductsOfUserController extends HttpServlet {
 
     private static final Injector INJECTOR = Injector.getInstance("vlad.petrovskyi.internetshop");
     private final ProductService productService =
@@ -19,16 +19,8 @@ public class AddProductController extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        req.getRequestDispatcher("/WEB-INF/views/products/addNewProduct.jsp").forward(req, resp);
-    }
-
-    @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-        String name = req.getParameter("name");
-        String price = req.getParameter("price");
-
-        productService.create(new Product(name, BigDecimal.valueOf(Long.parseLong(price))));
-        resp.sendRedirect(req.getContextPath() + "/allProductsFromDb");
+        List<Product> productList = productService.getAll();
+        req.setAttribute("products", productList);
+        req.getRequestDispatcher("WEB-INF/views/products/allProductsBuy.jsp").forward(req, resp);
     }
 }
