+++ src/main/java/vlad/petrovskyi/internetshop/controllers/user/DeleteUserController.java
@@ -1,4 +1,4 @@
-package vlad.petrovskyi.internetshop.controllers.db.orders;
+package vlad.petrovskyi.internetshop.controllers.user;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -6,16 +6,19 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import vlad.petrovskyi.internetshop.lib.Injector;
-import vlad.petrovskyi.internetshop.service.OrderService;
+import vlad.petrovskyi.internetshop.service.UserService;
 
-public class AllOrdersController extends HttpServlet {
+public class DeleteUserController extends HttpServlet {
+
     private static final Injector INJECTOR = Injector.getInstance("vlad.petrovskyi.internetshop");
-    private final OrderService orderService =
-            (OrderService) INJECTOR.getInstance(OrderService.class);
+    private final UserService userService = (UserService) INJECTOR.getInstance(UserService.class);
 
+    @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        req.setAttribute("orders", orderService.getAll());
-        req.getRequestDispatcher("/WEB-INF/views/orders/allOrders.jsp").forward(req, resp);
+        String userId = req.getParameter("user_id");
+        userService.delete(Long.valueOf(userId));
+
+        resp.sendRedirect(req.getContextPath() + "/getAllUsers");
     }
 }
