+++ src/main/webapp/web.xml
@@ -23,7 +23,7 @@
 
     <servlet>
         <servlet-name>getAllUsers</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.db.users.GetAllUsersController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.user.GetAllUsersController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>getAllUsers</servlet-name>
@@ -32,7 +32,7 @@
 
     <servlet>
         <servlet-name>deleteUser</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.db.users.DeleteUserController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.user.DeleteUserController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>deleteUser</servlet-name>
@@ -41,7 +41,7 @@
 
     <servlet>
         <servlet-name>getAllProductsFromDb</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.db.products.GetAllProductsController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.products.GetAllProductsInDbController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>getAllProductsFromDb</servlet-name>
@@ -50,7 +50,7 @@
 
     <servlet>
         <servlet-name>addNewProduct</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.db.products.AddProductController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.products.AddProductController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>addNewProduct</servlet-name>
@@ -59,7 +59,7 @@
 
     <servlet>
         <servlet-name>deleteProductFromDb</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.db.products.DeleteProductController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.products.DeleteProductController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>deleteProductFromDb</servlet-name>
@@ -77,7 +77,7 @@
 
     <servlet>
         <servlet-name>allProductsToBuy</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.user.GetAllProductsController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.products.GetAllProductsOfUserController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>allProductsToBuy</servlet-name>
@@ -113,7 +113,7 @@
 
     <servlet>
         <servlet-name>getAllOrders</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.db.orders.AllOrdersController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.orders.AllOrdersController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>getAllOrders</servlet-name>
@@ -122,7 +122,7 @@
 
     <servlet>
         <servlet-name>completeOrder</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.user.CompleteOrderController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.orders.CompleteOrderController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>completeOrder</servlet-name>
@@ -131,7 +131,7 @@
 
     <servlet>
         <servlet-name>userOrders</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.user.UserOrdersController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.orders.UserOrdersController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>userOrders</servlet-name>
@@ -149,7 +149,7 @@
 
     <servlet>
         <servlet-name>viewOrder</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.db.orders.ViewOrderController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.orders.ViewOrderController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>viewOrder</servlet-name>
@@ -158,7 +158,7 @@
 
     <servlet>
         <servlet-name>deleteOrder</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.db.orders.DeleteOrderController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.orders.DeleteOrderController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>deleteOrder</servlet-name>
@@ -167,7 +167,7 @@
 
     <servlet>
         <servlet-name>loginPage</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.user.LoginController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.LoginController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>loginPage</servlet-name>
